moves every view function into a separate file
